File restructuring to match suggested Meteor file structure
