+++ .gitignore
@@ -1,1 +1,2 @@
 packages/
+.DS_Store
