+++ .meteor/.finished-upgraders
@@ -15,3 +15,4 @@ notices-for-facebook-graph-api-2
 1.2.0-meteor-platform-split
 1.2.0-cordova-changes
 1.2.0-breaking-changes
+1.3.0-split-minifiers-package
