+++ .meteor/.gitignore
@@ -1,1 +1,2 @@
+dev_bundle
 local
