+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.2.0.2
+METEOR@1.3.4.4
