+++ .meteor/versions
@@ -1,79 +1,87 @@
-accounts-base@1.2.1
-autoupdate@1.2.3
-babel-compiler@5.8.24_1
-babel-runtime@0.1.4
-base64@1.0.4
-binary-heap@1.0.4
-blaze@2.1.3
-blaze-html-templates@1.0.1
-blaze-tools@1.0.4
-boilerplate-generator@1.0.4
-caching-compiler@1.0.0
-caching-html-compiler@1.0.2
-callback-hook@1.0.4
-check@1.0.6
-coffeescript@1.0.10
-ddp@1.2.2
-ddp-client@1.2.1
-ddp-common@1.2.1
-ddp-rate-limiter@1.0.0
-ddp-server@1.2.1
-deps@1.0.9
-diff-sequence@1.0.1
-easy:search@2.0.0
-easysearch:autosuggest@2.0.0
-easysearch:components@2.0.0
-easysearch:core@2.0.0
-easysearch:elasticsearch@2.0.0
-ecmascript@0.1.5
-ecmascript-collections@0.1.6
-ejson@1.0.7
-erasaur:meteor-lodash@3.10.1_1
-fastclick@1.0.7
-geojson-utils@1.0.4
-hot-code-push@1.0.0
-html-tools@1.0.5
-htmljs@1.0.5
-http@1.1.1
-id-map@1.0.4
-jeremy:selectize@0.1.3
-jquery@1.11.4
-launch-screen@1.0.4
-livedata@1.0.15
-localstorage@1.0.5
-logging@1.0.8
-meteor@1.1.9
-meteor-base@1.0.1
-minifiers@1.1.7
-minimongo@1.0.10
-mobile-experience@1.0.1
-mobile-status-bar@1.0.6
-mongo@1.1.2
-mongo-id@1.0.1
-npm-mongo@1.4.39_1
-observe-sequence@1.0.7
-ordered-dict@1.0.4
+accounts-base@1.2.8
+allow-deny@1.0.5
+autoupdate@1.2.11
+babel-compiler@6.8.4
+babel-runtime@0.1.9_1
+base64@1.0.9
+binary-heap@1.0.9
+blaze@2.1.8
+blaze-html-templates@1.0.4
+blaze-tools@1.0.9
+boilerplate-generator@1.0.9
+caching-compiler@1.0.6
+caching-html-compiler@1.0.6
+callback-hook@1.0.9
+check@1.2.3
+coffeescript@1.1.3
+ddp@1.2.5
+ddp-client@1.2.9
+ddp-common@1.2.6
+ddp-rate-limiter@1.0.5
+ddp-server@1.2.9
+deps@1.0.12
+diff-sequence@1.0.6
+easy:search@2.0.9
+easysearch:autosuggest@2.0.10
+easysearch:components@2.0.9
+easysearch:core@2.0.9
+easysearch:elasticsearch@2.0.9
+ecmascript@0.4.7
+ecmascript-runtime@0.2.12
+ejson@1.0.12
+erasaur:meteor-lodash@4.0.0
+fastclick@1.0.12
+geojson-utils@1.0.9
+hot-code-push@1.0.4
+html-tools@1.0.10
+htmljs@1.0.10
+http@1.1.8
+id-map@1.0.8
+jeremy:selectize@0.12.1_5
+jquery@1.11.9
+launch-screen@1.0.12
+livedata@1.0.18
+localstorage@1.0.11
+logging@1.0.14
+meteor@1.1.16
+meteor-base@1.0.4
+minifier-css@1.1.13
+minifier-js@1.1.13
+minimongo@1.0.17
+mobile-experience@1.0.4
+mobile-status-bar@1.0.12
+modules@0.6.5
+modules-runtime@0.6.5
+mongo@1.1.9_1
+mongo-id@1.0.5
+npm-mongo@1.4.45
+observe-sequence@1.0.12
+ordered-dict@1.0.8
 peerlibrary:assert@0.2.5
-peerlibrary:base-component@0.10.0
-peerlibrary:blaze-components@0.13.0
-promise@0.5.0
-random@1.0.4
-rate-limit@1.0.0
-reactive-dict@1.1.2
-reactive-var@1.0.6
-reload@1.1.4
-retry@1.0.4
-routepolicy@1.0.6
-service-configuration@1.0.5
-session@1.1.1
-spacebars@1.0.7
-spacebars-compiler@1.0.7
-standard-minifiers@1.0.1
-templating@1.1.4
-templating-tools@1.0.0
-tracker@1.0.9
-ui@1.0.8
-underscore@1.0.4
-url@1.0.5
-webapp@1.2.2
-webapp-hashing@1.0.5
+peerlibrary:base-component@0.14.0
+peerlibrary:blaze-components@0.16.2
+peerlibrary:computed-field@0.3.1
+peerlibrary:data-lookup@0.1.0
+peerlibrary:reactive-field@0.1.0
+promise@0.7.3
+random@1.0.10
+rate-limit@1.0.5
+reactive-dict@1.1.8
+reactive-var@1.0.10
+reload@1.1.10
+retry@1.0.8
+routepolicy@1.0.11
+service-configuration@1.0.10
+session@1.1.6
+spacebars@1.0.12
+spacebars-compiler@1.0.12
+standard-minifier-css@1.0.8
+standard-minifier-js@1.0.8
+templating@1.1.13
+templating-tools@1.0.4
+tracker@1.0.14
+ui@1.0.11
+underscore@1.0.9
+url@1.0.10
+webapp@1.2.10
+webapp-hashing@1.0.9
