+++ client/main.html
@@ -1,0 +1,14 @@
+<head>
+  <title>Leaderboard</title>
+  <meta name="viewport" content="width=device-width, user-scalable=no">
+  <link href='http://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600' rel='stylesheet' type='text/css'>
+</head>
+
+<body>
+    <div class="outer">
+      <div class="logo"></div>
+      <h1 class="title">Leaderboard</h1>
+      <div class="subtitle">Select a scientist to give them points</div>
+      {{> leaderboard}}
+    </div>
+</body>
