+++ client/main.js
@@ -1,0 +1,1 @@
+import '/imports/startup/client';
