+++ imports/api/players/players.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Players = new Mongo.Collection("players");
