+++ imports/startup/client/index.js
@@ -1,0 +1,2 @@
+import './autorun.js';
+import '../../ui/components/leaderboard/leaderboard.js';
