+++ imports/startup/server/index.js
@@ -1,0 +1,2 @@
+import './fixtures.js';
+import './register_apis.js';
