+++ imports/startup/server/register_apis.js
@@ -1,0 +1,1 @@
+import '../../api/players/players_index.js';
