+++ imports/ui/components/leaderboard/leaderboard.html
@@ -1,18 +1,3 @@
-<head>
-  <title>Leaderboard</title>
-  <meta name="viewport" content="width=device-width, user-scalable=no">
-  <link href='http://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600' rel='stylesheet' type='text/css'>
-</head>
-
-<body>
-    <div class="outer">
-      <div class="logo"></div>
-      <h1 class="title">Leaderboard</h1>
-      <div class="subtitle">Select a scientist to give them points</div>
-      {{> leaderboard}}
-    </div>
-</body>
-
 <template name="leaderboard">
   {{> EasySearch.Input index=index attributes=inputAttributes }}
 
@@ -68,13 +53,4 @@
     </ul>
   {{/if}}
 
-</template>
-
-<template name="player">
-  <li class="player {{selected}}">
-    <span class="name">{{name}}</span>
-    <span class="metadata">
-      <span class="category">{{category}}</span><span class="score">{{score}}</span>
-    </span>
-  </li>
 </template>
