+++ imports/ui/components/players/player.html
@@ -1,0 +1,8 @@
+<template name="player">
+  <li class="player {{selected}}">
+    <span class="name">{{name}}</span>
+    <span class="metadata">
+      <span class="category">{{category}}</span><span class="score">{{score}}</span>
+    </span>
+  </li>
+</template>
