+++ server/leaderboard.js
@@ -1,0 +1,10 @@
+import { Meteor } from 'meteor/meteor';
+import { Players } from '../imports/api/players/players.js';
+import { check } from 'meteor/check';
+
+Meteor.methods({
+  updateScore: function (playerId) {
+    check(playerId, String);
+    Players.update(playerId, { $inc: { score: 5 }});
+  }
+});
