+++ server/main.js
@@ -1,0 +1,2 @@
+import './leaderboard.js';
+import '/imports/startup/server';
